feat(d8x): track fees in a dedicated adapter with per-pool pricing
